implement roles and permissions to edit and delete features and comments

+++ app/Http/Controllers/CommentController.php
@@ -33,6 +33,9 @@ class CommentController extends Controller
     }
     public function destroy(Comment $comment)
     {
+        if ($comment->user_id !== auth()->id()) {
+            abort(403);
+        }
         $comment->delete();
         return back();
     }
